MCVE for "Called stop() outside of a test context" error
When a user signs in through the development environment with email
address "user@example.com" and password "validPassword", it will
succeed in the development environment and transition to the "projects"
route. When attempting to execute an acceptance test performing the same
steps, however, the test will fail in Phantom JS, saying that the
currentURL remains "/singin" when "/projects" was expected. If the test
is run inside of Safari or Chrome, however, there will be an additional
JSHint error message saying "Uncaught Error: Called stop() outside of a
test context."